desenv

+++ www/js/controllers/ListUsuariosCtrl.js
@@ -1,5 +1,5 @@
 angular.module('app.controllers')
-.controller('ListUsuariosCtrl', function($scope, $http, $window, DatabaseValues, $state) {
+.controller('ListUsuariosCtrl', function($scope, $http, $window, DatabaseValues, $state, $ionicPopup) {
 	
 	$scope.usuarios = [];
 
@@ -17,14 +17,30 @@ angular.module('app.controllers')
 	
 	$scope.removeUsuario = function(usuarioId) {
 		
-		DatabaseValues.setup();
-		DatabaseValues.bancoDeDados.transaction(function(transacao) {
-			transacao.executeSql('DELETE FROM usuario WHERE id = ?', [usuarioId]);
+		var confirmPopup = $ionicPopup.confirm({
+			
+			title: 'Remover Usuário',
+			template: 'Tem certeza que deseja remover o usuário?',
+			cancelText: 'Cancelar',
+			okText: 'Confirmar'
+			
+		}).then(function(res) {
+			if (res) {
+				
+				DatabaseValues.setup();
+				DatabaseValues.bancoDeDados.transaction(function(transacao) {
+					transacao.executeSql('DELETE FROM usuario WHERE id = ?', [usuarioId]);
+				});
+			  
+				// RELOAD AQUI
+				
+				//$state.go($state.current, {}, {reload: true}); 
+				$window.location.reload(true);				
+				
+				console.log('confirmed');
+				
+			}
 		});
-	  
-		// RELOAD AQUI
-		
-		$state.go($state.current, {}, {reload: true}); 
 	  
     };
